BrotliCompressor.Free, finalizer, use q=9 for enc tests
