+++ enc/encode_test.go
@@ -1,33 +1,86 @@
 package enc
 
 import (
+	"bytes"
+	"log"
 	"strings"
 	"testing"
 )
 
+const (
+	testQuality int = 9
+)
+
 func TestBufferSizes(T *testing.T) {
+	params := NewBrotliParams()
+	params.SetQuality(testQuality)
 
 	input1 := []byte(strings.Repeat("The quick brown fox jumps over the lazy dog", 100000))
+	log.Printf("q=%d, inputSize=%d\n", params.Quality(), len(input1))
+
 	output1 := make([]byte, len(input1)*2)
-	_, err := CompressBuffer(nil, input1, output1)
+	_, err := CompressBuffer(params, input1, output1)
 	if err != nil {
 		T.Error(err)
 	}
 
 	output2 := make([]byte, len(input1))
-	_, err = CompressBuffer(nil, input1, output2)
+	_, err = CompressBuffer(params, input1, output2)
 	if err != nil {
 		T.Error(err)
 	}
 
 	output3 := make([]byte, len(input1)/2)
-	_, err = CompressBuffer(nil, input1, output3)
+	_, err = CompressBuffer(params, input1, output3)
 	if err != nil {
 		T.Error(err)
 	}
 
-	_, err = CompressBuffer(nil, input1, nil)
+	_, err = CompressBuffer(params, input1, nil)
 	if err != nil {
 		T.Error(err)
+	}
+}
+
+func TestStreamEncode(T *testing.T) {
+	params := NewBrotliParams()
+	params.SetQuality(testQuality)
+
+	input1 := []byte(strings.Repeat("The quick brown fox jumps over the lazy dog", 100000))
+	inputSize := int64(len(input1))
+	log.Printf("q=%d, inputSize=%d\n", params.Quality(), inputSize)
+
+	for lgwin := 16; lgwin <= 22; lgwin += 1 {
+		params.SetLgwin(lgwin)
+		compressor := NewBrotliCompressor(params)
+		defer compressor.Free()
+		blockSize := compressor.GetInputBlockSize()
+
+		fullOutput := make([]byte, 0)
+
+		rounds := 0
+		pos := int64(0)
+		for pos < inputSize {
+			rounds++
+			copySize := blockSize
+			remaining := inputSize - pos
+			if remaining < copySize {
+				copySize = remaining
+			}
+			compressor.CopyInputToRingBuffer(input1[pos : pos+copySize])
+			pos += copySize
+
+			output := compressor.WriteBrotliData(pos >= inputSize, false)
+			fullOutput = append(fullOutput, output...)
+		}
+
+		fullOutput2, _ := CompressBuffer(params, input1, make([]byte, 0))
+
+		if !bytes.Equal(fullOutput, fullOutput2) {
+			T.Fatal("for lgwin %d, stream compression didn't give same result as buffer compression", params.Lgwin())
+		}
+
+		outputSize := len(fullOutput)
+		log.Printf("lgwin=%d, rounds=%d, output=%d (%.4f%% of input size)\n", params.Lgwin(), rounds, outputSize, float32(outputSize)*100.0/float32(inputSize))
 	}
 }
